LIMS-2489: Localise images before passing HTML to WeasyPrint.
WeasyPrint will attempt to retrieve attachments directly from the URL
referenced in the HTML report, which may refer back to a single-threaded
(and currently occupied) zeoclient, hanging it.  All "attachments"
using urls ending with at_download/AttachmentFile must be converted
to local files.
